Got the following error before the weekend (monday as of this commit)
Traceback (most recent call last):
  File "regionTestingFunctions.py", line 1052, in <module>
    regionTestingFunc("BP2", "ppXSHSM", BP2_dictPointlistAtlas[2], "angle", "plottingLimits/Atlas2023/BP2_Atlas", "BP2_Atlas2023_obs_limit", logyscale = True, individualPlots = True)
  File "regionTestingFunctions.py", line 703, in regionTestingFunc
    for dictElement in userParametersDict:
RuntimeError: dictionary changed size during iteration

Do not know why and it fixed itself today when trying to replicate the bug (same seed as can be seen in the .ini file). We do this commit if for any
reason we try find how to replicate the bug and fix the bug.