Added convertJSON2TSV, to convert the json file to a csv file for easy access of data in parameterData.py, however, the last decimals (~10^-14) is different compared to the json file. I believe this is just the python float data types are different. Note that anyway you use pandas will give you the same floats (even if you do not convert to a csv file and direct access the data through pandas). So you can ignore this
The full data is found in Atlas2023Limits.tsv and BP2 constrained in the appropriate file.

parameterData.py is nearly done. Only mkdir is behaving weird. I believe this is due to concurrency
issues with multiprocessing. Should be fixed if you make the targetDir in mprocMain.
You do not have to modify the other functions.